# ⚡ Faster `Hex` for netstandard2.1 and up